requirements.txt
opencv-python
numpy
matplotlib
PyYAML
Pillow
requests
torch
torchvision
tqdm
pandas
av
ptflops
yacs
